pyproject.toml
[tool.poetry]
name = "anverali-test-task"
version = "0.1.0"
description = ""
authors = ["disqurell <lisyachenko.daniil@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
flask = "^3.0.3"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
